Refactor: Container 컴포넌트 분리
