+++ frontend/src/components/Container/index.js
@@ -1,0 +1,23 @@
+import styled from 'styled-components';
+
+const Container = styled.div`
+  @font-face {
+    font-family: 'TmoneyRoundWindRegular';
+    src: url('https://cdn.jsdelivr.net/gh/projectnoonnu/noonfonts_20-07@1.0/TmoneyRoundWindRegular.woff')
+      format('woff');
+    font-weight: normal;
+    font-style: normal;
+  }
+  width: 16rem;
+  height: 100vh;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  /* text-align: center; */
+  margin: 0 auto;
+  font-family: 'TmoneyRoundWindRegular';
+  font-weight: 400;
+`;
+
+export default Container;
